+++ tools/mkdocs_macros_bids/macros.py
@@ -2,7 +2,7 @@
 import os
 import sys
 
-from bidsschematools import render, schema, utils
+from bidsschematools import render, schema
 
 code_path = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 sys.path.append(code_path)
@@ -61,12 +61,17 @@ def _get_source_path(level=1):
     return caller.f_locals["_Context__self"]["page"].file.src_path
 
 
-def make_filename_template(**kwargs):
-    """Generate a filename template snippet from the schema, based on specific
-    filters.
+def make_filename_template(src_path=None, pdf_format=False, **kwargs):
+    """Generate a filename template snippet from the schema, based on specific filters.
 
     Parameters
     ----------
+    pdf_format : bool, optional
+        If True, the filename template will be compiled as a standard markdown code block,
+        without any hyperlinks, so that the specification's PDF build will look right.
+        If False, the filename template will use HTML and include hyperlinks.
+        This works on the website.
+        Default is False.
     kwargs : dict
         Keyword arguments used to filter the schema.
         Example kwargs that may be used include: "suffixes", "datatypes",
@@ -78,13 +83,20 @@ def make_filename_template(**kwargs):
         A multiline string containing the filename templates for file types
         in the schema, after filtering.
     """
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
-    codeblock = render.make_filename_template(schema_obj, **kwargs)
+    if src_path is None:
+        src_path = _get_source_path()
+
+    schema_obj = schema.load_schema()
+    codeblock = render.make_filename_template(
+        schema_obj,
+        src_path=src_path,
+        pdf_format=pdf_format,
+        **kwargs,
+    )
     return codeblock
 
 
-def make_entity_table(**kwargs):
+def make_entity_table(src_path=None, **kwargs):
     """Generate an entity table from the schema, based on specific filters.
 
     Parameters
@@ -100,9 +112,11 @@ def make_entity_table(**kwargs):
         A Markdown-format table containing the corresponding entity table for
         a subset of the schema.
     """
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
-    table = render.make_entity_table(schema_obj, **kwargs)
+    if src_path is None:
+        src_path = _get_source_path()
+
+    schema_obj = schema.load_schema()
+    table = render.make_entity_table(schema_obj, src_path=src_path, **kwargs)
     return table
 
 
@@ -118,8 +132,8 @@ def make_entity_definitions(src_path=None):
     """
     if src_path is None:
         src_path = _get_source_path()
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
+
+    schema_obj = schema.load_schema()
     text = render.make_entity_definitions(schema_obj, src_path=src_path)
     return text
 
@@ -141,8 +155,8 @@ def make_glossary(src_path=None):
     """
     if src_path is None:
         src_path = _get_source_path()
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
+
+    schema_obj = schema.load_schema()
     text = render.make_glossary(schema_obj, src_path=src_path)
     return text
 
@@ -166,8 +180,8 @@ def make_suffix_table(suffixes, src_path=None):
     """
     if src_path is None:
         src_path = _get_source_path()
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
+
+    schema_obj = schema.load_schema()
     table = render.make_suffix_table(schema_obj, suffixes, src_path=src_path)
     return table
 
@@ -196,9 +210,34 @@ def make_metadata_table(field_info, src_path=None):
     """
     if src_path is None:
         src_path = _get_source_path()
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
+
+    schema_obj = schema.load_schema()
     table = render.make_metadata_table(schema_obj, field_info, src_path=src_path)
+    return table
+
+
+def make_sidecar_table(table_name, src_path=None):
+    """Generate a markdown table of metadata field information.
+
+    Parameters
+    ----------
+    table_name : str or list of str
+        Qualified name(s) in schema.rules.sidecars
+    src_path : str | None
+        The file where this macro is called, which may be explicitly provided
+        by the "page.file.src_path" variable.
+
+    Returns
+    -------
+    table : str
+        A Markdown-format table containing the corresponding table for
+        the requested fields.
+    """
+    if src_path is None:
+        src_path = _get_source_path()
+
+    schema_obj = schema.load_schema()
+    table = render.make_sidecar_table(schema_obj, table_name, src_path=src_path)
     return table
 
 
@@ -229,10 +268,12 @@ def make_subobject_table(object_tuple, field_info, src_path=None):
     if src_path is None:
         src_path = _get_source_path()
 
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
+    schema_obj = schema.load_schema()
     table = render.make_subobject_table(
-        schema_obj, object_tuple, field_info, src_path=src_path
+        schema_obj,
+        object_tuple,
+        field_info,
+        src_path=src_path,
     )
     return table
 
@@ -261,8 +302,8 @@ def make_columns_table(column_info, src_path=None):
     """
     if src_path is None:
         src_path = _get_source_path()
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
+
+    schema_obj = schema.load_schema()
     table = render.make_columns_table(schema_obj, column_info, src_path=src_path)
     return table
 
@@ -303,7 +344,6 @@ def define_common_principles(src_path=None):
     if src_path is None:
         src_path = _get_source_path()
 
-    schemapath = utils.get_schema_path()
-    schema_obj = schema.load_schema(schemapath)
+    schema_obj = schema.load_schema()
     string = render.define_common_principles(schema_obj, src_path=src_path)
     return string